Supporting perf events with opentracing
Platform counters provide huge insights into what's happening within a
span. Although, latency of a span can help us in identifying the bottle
necks within a trace, perfevents can compliment that in identifying as
to where was the bottle neck and which platform resource was busy for
that span. This patch is in addition to the perf events support to
zipkin-go-opentracing.

Signed-off-by: Hemant Kumar <hemant@linux.vnet.ibm.com>